EJBCAINTER-72: Fixes after code review part 2
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@22845 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/cesecore-common/src/org/cesecore/util/SecurityFilterInputStream.java
@@ -37,7 +37,7 @@ public class SecurityFilterInputStream extends FilterInputStream{
         super(inputStream);
     }
     
-    public SecurityFilterInputStream(InputStream inputStream, int maxBytes){
+    public SecurityFilterInputStream(InputStream inputStream, long maxBytes){
         super(inputStream);
         this.maxBytes = maxBytes;
     }
